Redid most of the code in Game.java, made it look more A E S T H E T I C. Also, the first version with something looking like a greeting menu

+++ src/org/academiadecodigo/jungleweed/GameObjects/SGFXGrabTotemIndicator.java
@@ -23,6 +23,8 @@ public class SGFXGrabTotemIndicator extends AbstractIndicator {
 
     @Override
     public void hide() {
-        this.text.delete();
+        if (this.text != null) {
+            this.text.delete();
+        }
     }
 }
